requirements.txt
ccxt
pandas
numpy
tensorflow
ta
requests
transformers
torch
python-telegram-bot
fredapi
scikit-learn
flask
coingecko-sdk
twikit